refactor(tests,client): import re-exported symbols from canonical modules
Phase 4 demolition prerequisite. Migrates importer sites from
notebooklm._session to canonical source modules:

- DEFAULT_TIMEOUT, DEFAULT_KEEPALIVE_MIN_INTERVAL,
  DEFAULT_MAX_CONCURRENT_RPCS, DEFAULT_MAX_CONCURRENT_UPLOADS:
    client.py → _session_config
- MAX_RETRY_AFTER_SECONDS, _parse_retry_after, _AuthSnapshot,
  _TransportAuthExpired, _TransportRateLimited, _TransportServerError:
    tests/{unit,integration} → _authed_transport
- ERROR_INJECT_ENV_VAR, _get_error_injection_mode:
    tests/conftest.py, tests/unit/test_vcr_config.py,
    tests/unit/test_session_lifecycle.py → _error_injection
- is_auth_error:
    tests/unit/test_client.py,
    tests/integration/test_session_integration.py → _session_helpers
- _resolve_keepalive_interval:
    tests/unit/test_session_lifecycle.py → _session_helpers

Also deleted tests/unit/test_transport_drain.py::
test_token_reexported_from_core_module — the test specifically asserted
the legacy re-export anchor at notebooklm._session._TransportOperationToken
still works; with the anchor going away in Task 10.A.5, the test is
obsolete (and tautological — it would just re-prove the canonical home
imports its own symbol).

Session class itself stays imported from _session; only secondary
re-exports moved.

Phase 4 Task 10.A.3 of the v0.5.0 refactor program.
